+++ packages/gamut/src/GridForm/GridFormInputGroup/index.tsx
@@ -2,6 +2,7 @@ import styled from '@emotion/styled';
 import React from 'react';
 import { UseFormMethods } from 'react-hook-form';
 
+import { Box } from '../../Box';
 import { FormError, FormGroup, FormGroupLabel } from '../../Form';
 import { HiddenText } from '../../HiddenText';
 import { Column } from '../../Layout';
@@ -20,11 +21,12 @@ export type GridFormInputGroupProps = {
   field: GridFormField;
   register: UseFormMethods['register'];
   setValue: (value: any) => void;
+  required?: boolean;
+  showRequired?: boolean;
 };
 
 const StyledFormGroup = styled(FormGroup)`
   margin-bottom: 0;
-
   // This is always the input
   > *:last-child {
     width: 100%;
@@ -32,12 +34,15 @@ const StyledFormGroup = styled(FormGroup)`
 `;
 
 export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
-  field,
+  error,
   isFirstError,
+  field,
   register,
   setValue,
-  error,
+  showRequired,
 }) => {
+  const errorMessage = error || field.customError;
+
   const getInput = () => {
     switch (field.type) {
       case 'checkbox':
@@ -49,7 +54,7 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
             field={field}
             register={register}
             setValue={setValue}
-            error={error}
+            error={errorMessage}
           />
         );
 
@@ -65,7 +70,7 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'select':
         return (
           <GridFormSelectInput
-            error={!!error}
+            error={!!errorMessage}
             field={field}
             register={register}
           />
@@ -74,7 +79,7 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
       case 'file':
         return (
           <GridFormFileInput
-            error={!!error}
+            error={!!errorMessage}
             field={field}
             register={register}
           />
@@ -82,13 +87,17 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
 
       case 'textarea':
         return (
-          <GridFormTextArea error={!!error} field={field} register={register} />
+          <GridFormTextArea
+            error={!!errorMessage}
+            field={field}
+            register={register}
+          />
         );
 
       default:
         return (
           <GridFormTextInput
-            error={!!error}
+            error={!!errorMessage}
             field={field}
             register={register}
           />
@@ -108,15 +117,17 @@ export const GridFormInputGroup: React.FC<GridFormInputGroupProps> = ({
 
   return (
     <Column size={field.size}>
-      <StyledFormGroup>
-        {field.hideLabel ? <HiddenText>{label}</HiddenText> : label}
-        {error && (
-          <FormError aria-live={isFirstError ? 'assertive' : 'off'}>
-            {error}
-          </FormError>
-        )}
-        {getInput()}
-      </StyledFormGroup>
+      <Box>
+        <StyledFormGroup>
+          {field.hideLabel ? <HiddenText>{label}</HiddenText> : label}
+          {getInput()}
+          {errorMessage && (
+            <FormError aria-live={isFirstError ? 'assertive' : 'off'}>
+              {errorMessage}
+            </FormError>
+          )}
+        </StyledFormGroup>
+      </Box>
     </Column>
   );
 };